Dev branch

+++ src/edu/uiowa/cs/similarity/Similarity.java
@@ -5,7 +5,7 @@ import java.util.Iterator;
 public class Similarity {
     public static double getMagnitude(Vector v) {
         double magniutde = 0;
-        int square;
+        double square;
         Iterator<String> keys = v.getKeySet().iterator();
         String key;
         while (keys.hasNext()) {
@@ -13,7 +13,6 @@ public class Similarity {
             square = v.getSimValue(key);
             magniutde += square * square;
         }
-        //System.out.println("Magnitude of " + v.base + ": " + magniutde);
         return magniutde;
     }
 }
